@@ -2,7 +2,7 @@ package io.roach.product.config;
 
 import java.util.Properties;
 
-import javax.persistence.EntityManagerFactory;
+import jakarta.persistence.EntityManagerFactory;
 import javax.sql.DataSource;
 
 import org.hibernate.cache.internal.NoCachingRegionFactory;
