@@ -3,8 +3,8 @@ package io.roach.trading.domain.account;
 import java.util.Optional;
 import java.util.UUID;
 
-import javax.persistence.LockModeType;
-import javax.persistence.QueryHint;
+import jakarta.persistence.LockModeType;
+import jakarta.persistence.QueryHint;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
