@@ -2,16 +2,12 @@ package io.roach.trading.util;
 
 import java.sql.Types;
 
-import org.hibernate.dialect.CockroachDB201Dialect;
-import org.hibernate.dialect.function.SQLFunctionTemplate;
+import org.hibernate.dialect.CockroachDialect;
 import org.hibernate.dialect.identity.IdentityColumnSupport;
 import org.hibernate.dialect.identity.IdentityColumnSupportImpl;
-import org.hibernate.type.StandardBasicTypes;
 
-public class CockroachDBDialect extends CockroachDB201Dialect {
+public class CockroachDBDialect extends CockroachDialect {
     public CockroachDBDialect() {
-        registerFunction("cluster_logical_timestamp",
-                new SQLFunctionTemplate(StandardBasicTypes.TIMESTAMP, "cluster_logical_timestamp()"));
     }
 
     @Override
